reorganizing folder and files follow KUVA practice
