+++ src/domain/app/BrowserApp.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { I18nextProvider } from 'react-i18next';
 
+import i18n from '../../common/translation/i18n/i18nInit';
 import App from './App';
-import i18n from './i18n/i18nInit';
 
 function BrowserApp() {
   return (
